Exportmanager code refactoring
