+++ exp/services/ledgerexporter/internal/exportmanager.go
@@ -2,88 +2,15 @@ package exporter
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/stellar/go/ingest/ledgerbackend"
 	"github.com/stellar/go/support/errors"
 	"github.com/stellar/go/xdr"
 )
 
-// LedgerCloseMetaObject represents a file with metadata and binary data.
-type LedgerCloseMetaObject struct {
-	// file name
-	objectKey string
-	// Actual binary data
-	data xdr.LedgerCloseMetaBatch
-}
-
-func NewLedgerCloseMetaObject(key string, startSeq uint32, endSeq uint32) *LedgerCloseMetaObject {
-	return &LedgerCloseMetaObject{
-		objectKey: key,
-		data: xdr.LedgerCloseMetaBatch{
-			StartSequence: xdr.Uint32(startSeq),
-			EndSequence:   xdr.Uint32(endSeq),
-		},
-	}
-}
-
-func (f *LedgerCloseMetaObject) GetLastLedgerCloseMetaSequence() (uint32, error) {
-	if len(f.data.LedgerCloseMetas) == 0 {
-		return 0, errors.New("LedgerCloseMetas is empty")
-	}
-
-	return f.data.LedgerCloseMetas[len(f.data.LedgerCloseMetas)-1].LedgerSequence(), nil
-}
-
-// AddLedgerCloseMeta adds a ledger
-func (f *LedgerCloseMetaObject) AddLedgerCloseMeta(ledgerCloseMeta xdr.LedgerCloseMeta) error {
-	lastSequence, err := f.GetLastLedgerCloseMetaSequence()
-	if err == nil {
-		if ledgerCloseMeta.LedgerSequence() != lastSequence+1 {
-			return fmt.Errorf("ledgers must be added sequentially. Sequence number: %d, "+
-				"expected sequence number: %d", ledgerCloseMeta.LedgerSequence(), lastSequence+1)
-		}
-	}
-
-	f.data.LedgerCloseMetas = append(f.data.LedgerCloseMetas, ledgerCloseMeta)
-	return nil
-}
-
-// LedgerCount returns the number of ledgers added so far
-func (f *LedgerCloseMetaObject) LedgerCount() uint32 {
-	return uint32(len(f.data.LedgerCloseMetas))
-}
-
 type ExporterConfig struct {
 	LedgersPerFile    uint32 `toml:"ledgers_per_file"`
 	FilesPerPartition uint32 `toml:"files_per_partition"`
-}
-
-// getObjectKey generates the file name based on the ledger sequence.
-func (e *ExporterConfig) getObjectKey(ledgerSeq uint32) (string, error) {
-	var objectKey string
-
-	if e.FilesPerPartition > 1 {
-		partitionSize := e.LedgersPerFile * e.FilesPerPartition
-		partitionStart := (ledgerSeq / partitionSize) * partitionSize
-		partitionEnd := partitionStart + partitionSize - 1
-		objectKey = fmt.Sprintf("%v-%v/", partitionStart, partitionEnd)
-	}
-
-	if e.LedgersPerFile < 1 {
-		return "", errors.New("Ledgers per file must be at least 1")
-	}
-
-	fileStart := (ledgerSeq / e.LedgersPerFile) * e.LedgersPerFile
-	fileEnd := fileStart + e.LedgersPerFile - 1
-	objectKey += fmt.Sprintf("%v", fileStart)
-
-	// Multiple ledgers per file
-	if fileStart != fileEnd {
-		objectKey += fmt.Sprintf("-%v", fileEnd)
-	}
-
-	return objectKey, nil
 }
 
 // ExportManager manages the creation and handling of export objects.
@@ -119,7 +46,7 @@ func (e *exportManager) AddLedgerCloseMeta(ledgerCloseMeta xdr.LedgerCloseMeta) 
 	ledgerSeq := ledgerCloseMeta.LedgerSequence()
 
 	// Determine the object key for the given ledger sequence
-	objectKey, err := e.config.getObjectKey(ledgerSeq)
+	objectKey, err := GetObjectKeyFromSequenceNumber(e.config, ledgerSeq)
 	if err != nil {
 		return errors.Wrapf(err, "failed to get object key for ledger %d", ledgerSeq)
 	}
